+++ jkit-sql/src/test/java/com/alianga/jkit/sql/SqlGoldenCorpusTest.java
@@ -45,6 +45,9 @@ public class SqlGoldenCorpusTest {
                 {"mysql", "SELECT 1"},
                 {"mysql", "SELECT 1 FROM dual"},
                 {"mysql", "SELECT * FROM t"},
+                {"mysql", "SELECT * FROM t PARTITION (p0, p1) WHERE id > 0"},
+                {"mysql", "SELECT a FROM emp PARTITION (p2020) e WHERE e.id = 1"},
+                
                 {"mysql", "SELECT t.* FROM t"},
                 {"mysql", "SELECT DISTINCT a, b FROM t"},
                 {"mysql", "SELECT a AS x, b y FROM t"},